[FIR] Make base checkers invariant
This prevents a situation like when a checker extends `FirCallChecker`
but is only added to `functionCallCheckers` and is therefore not called
on annotation calls.

Two such cases were discovered, namely `FirSuspendModifierChecker` and
`FirNamedVarargChecker`, however, they were added to the general and the
specific set of checkers; thus no bug was present, only some redundancy.